1.simple-lieve-demo/requirements.txt
google-genai
python-dotenv
numpy
wave
sounddevice